+++ changelog.md
@@ -1,13 +1,16 @@
 # Changelog
 
-## [1.10.3] - 2026-1-16 Surveys
+## [1.10.3] - 2026-1-21 Surveys
 
 ### New Features
 - Surveys (Multiple Choice, Multiple Selection, Open Ended, Yes/No, Scale, Ranking, Availability, Time, Date)
+- Public, Private, and Discord Guild locked Surveys
+- Survey History to view your responses and past surveys
 
 ### Improvements
 - Changelog page now selects the most recent update or scrolls to the selected update
 - Discord Guild invite requests now display average response time.
+- Leaderboard filter updates url for easier sharing
 
 ### Bug Fixes
 - Clicking on a link to a Law now respects the user's Reduce Motion preferences